Declare strbuffer static to move them into BSS and off stacks in functions. Fix typo in disable function management. Fix declarations of q2admin_realloc and q2admin_free to align with current C. Declare maxspawn_cmds size_t.

+++ zb_spawn.c
@@ -47,7 +47,7 @@ spawncmd_t;
 #define SPAWN_RE  2
 
 spawncmd_t spawncmds[SPAWN_MAXCMDS];
-long long maxspawn_cmds = 0;
+size_t maxspawn_cmds = 0;
